refactor(node): move node to its own crate
make the interledger crate only the library that bundles the main functionality together
