+++ crates/interledger-node/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "interledger-node"
+version = "0.4.1-alpha.1"
+authors = ["Evan Schwartz <evan@ripple.com>"]
+description = "Interledger node (sender, connector, receiver bundle)"
+license = "Apache-2.0"
+edition = "2018"
+repository = "https://github.com/interledger-rs/interledger-rs"
+
+[dependencies]
+bytes = "0.4.12"
+clap = "2.32.0"
+config = "0.9.3"
+env_logger = "0.6.1"
+futures = "0.1.25"
+hex = "0.3.2"
+interledger = { path = "../interledger", version = "^0.4.1-alpha.1" }
+lazy_static = "1.4.0"
+log = "0.4.6"
+ring = "0.14.6"
+serde = "1.0.99"
+tokio = "0.1.20"
+url = "2.1.0"
+libc = "0.2.62"
+warp = "0.1.20"
+secrecy = { version = "0.4.0", features = ["serde"] }
+
+[dev-dependencies]
+net2 = "0.2.33"
+rand = "0.6.5"
+redis = "0.12.0"
+reqwest = "0.9.16"
+serde_json = "1.0.39"
+secrecy = "0.4.0"
+
+[badges]
+circle-ci = { repository = "interledger-rs/interledger-rs" }
+codecov = { repository = "interledger-rs/interledger-rs" }
