+++ crates/interledger-node/README.md
@@ -1,0 +1,6 @@
+# Interledger.rs
+
+This crate provides both a command-line tool and a library API for the Rust
+implementation of [the Interledger Protocol Suite](https://interledger.org/rfcs/0001-interledger-architecture/);
+your one-stop shop for developers looking to work with Interledger.
+It is the ultimate consumer of the other crates in this repository.
