+++ crates/interledger-node/src/lib.rs
@@ -1,0 +1,2 @@
+mod node;
+pub use node::*;
