+++ crates/interledger-service-util/src/exchange_rates_service.rs
@@ -472,5 +472,4 @@ mod tests {
             handler,
         )
     }
-
 }
