+++ crates/interledger-settlement-engines/Cargo.toml
@@ -55,8 +55,7 @@ net2 = "0.2.33"
 os_type = "2.2.0"
 rand = "0.7.0"
 interledger = { path = "../interledger", version = "^0.4.1-alpha.1"}
-interledger-packet = { path = "../interledger-packet", version = "^0.2.2-alpha.1"}
-interledger-store-redis = { path = "../interledger-store-redis", version = "^0.2.2-alpha.1"}
+interledger-node = { path = "../interledger-node", version = "^0.4.1-alpha.1"}
 
 [features]
 ethereum = ["tiny-keccak", "secp256k1", "rlp", "ethereum-types"]
