+++ crates/interledger-settlement-engines/tests/eth_ledger_settlement.rs
@@ -4,9 +4,8 @@
 use env_logger;
 use futures::future::join_all;
 use futures::Future;
-use interledger::node::{random_secret, AccountDetails, InterledgerNode};
-use interledger_packet::Address;
-use interledger_service::Username;
+use interledger::{api::AccountDetails, packet::Address, service::Username};
+use interledger_node::{random_secret, InterledgerNode};
 use secrecy::{ExposeSecret, SecretBytes, SecretString};
 use std::net::SocketAddr;
 use std::str::FromStr;
