+++ crates/interledger-settlement-engines/tests/xrp_ledger_settlement.rs
@@ -2,9 +2,8 @@
 
 use futures::future::join_all;
 use futures::Future;
-use interledger::node::{random_secret, AccountDetails, InterledgerNode};
-use interledger_packet::Address;
-use interledger_service::Username;
+use interledger::{api::AccountDetails, packet::Address, service::Username};
+use interledger_node::{random_secret, InterledgerNode};
 use secrecy::SecretString;
 use std::str::FromStr;
 use tokio::runtime::Builder as RuntimeBuilder;
