+++ crates/interledger-settlement/src/message_service.rs
@@ -243,5 +243,4 @@ mod tests {
             .as_bytes(),
         );
     }
-
 }
